+++ src/www/ui_components/app-root.js
@@ -47,7 +47,7 @@ import './outline-icons.js';
 import './privacy-view.js';
 import './server_list.ts';
 import './server_card.ts';
-import './server_connection_viz.ts';
+import './server_connection_viz';
 import './servers-view.js';
 import './server-rename-dialog.js';
 import './user-comms-dialog.js';
